terraform/src/cloud_functions/concatenate_wav_translate_m4a/requirements.txt
google-auth==2.16.2
requests==2.27.1
google-cloud-storage==2.0.0
pydub==0.25.1